refactor LineageTab
